frontend/requirements.txt
# Streamlit requirements
streamlit
requests
uuid
python-dotenv
msal-streamlit-auth
